fix fmt CI, move unit test injection_queue_depth to stable
